controllers/crate/version: Move google.visualization code into component
