+++ app/controllers/crate/version.js
@@ -165,44 +165,7 @@ export default Ember.Controller.extend({
                 data.push(row);
             }
 
-            // TODO: move this to a component
-            function drawChart() {
-                if (!window.google || !window.googleChartsLoaded) {
-                    Ember.$('.graph').hide();
-                    return;
-                } else {
-                    Ember.$('.graph').show();
-                }
-                var myData = window.google.visualization.arrayToDataTable(data);
-
-                var fmt = new window.google.visualization.DateFormat({
-                    pattern: 'LLL d, yyyy',
-                });
-                fmt.format(myData, 0);
-                var el = document.getElementById('graph-data');
-                if (!el) {
-                    return;
-                }
-                var chart = new window.google.visualization.AreaChart(el);
-                chart.draw(myData, {
-                    chartArea: { 'left': 85, 'width': '77%', 'height': '80%' },
-                    hAxis: {
-                        minorGridlines: { count: 8 },
-                    },
-                    vAxis: {
-                        minorGridlines: { count: 5 },
-                        viewWindow: { min: 0, },
-                    },
-                    isStacked: true,
-                    focusTarget: 'category',
-                });
-            }
-
-            Ember.run.scheduleOnce('afterRender', this, drawChart);
-            Ember.$(window).off('resize.chart');
-            Ember.$(window).on('resize.chart', drawChart);
-            Ember.$(document).off('googleChartsLoaded');
-            Ember.$(document).on('googleChartsLoaded', drawChart);
+            this.set('downloadData', data);
         },
     },
 });
