+++ app/templates/crate/version.hbs
@@ -231,7 +231,7 @@
     </div>
     <div class='graph'>
         <h4>Downloads over the last 90 days</h4>
-        {{download-graph}}
+        {{download-graph data=downloadData}}
     </div>
 </div>
 {{/if}}
